+++ apps/reactjs/src/routes/__root.tsx
@@ -1,0 +1,13 @@
+import { Outlet, createRootRoute } from '@tanstack/react-router';
+
+import '../styles.css';
+
+export const Route = createRootRoute({ component: RootComponent });
+
+function RootComponent() {
+  return (
+    <>
+      <Outlet />
+    </>
+  );
+}