+++ apps/reactjs/vite.config.ts
@@ -1,0 +1,15 @@
+import { defineConfig } from 'vite';
+
+import { tanstackRouter } from '@tanstack/router-plugin/vite';
+
+import viteReact from '@vitejs/plugin-react';
+
+const config = defineConfig({
+  resolve: { tsconfigPaths: true },
+  plugins: [
+    tanstackRouter({ target: 'react', autoCodeSplitting: true }),
+    viteReact(),
+  ],
+});
+
+export default config;