+++ packages/reactjs/src/index.ts
@@ -1,0 +1,2 @@
+export * from './useSubscriber';
+export * as prime from '@bemedev/subscriber';