+++ packages/solidjs/src/index.ts
@@ -1,0 +1,2 @@
+export * from './createSubscriber';
+export * as prime from '@bemedev/subscriber';